build: centralize ktfmt conventions in build-logic; fix deprecations
Move the shared ktfmt + dependency-hygiene setup out of the root
`allprojects {}` block into a `harc.base-conventions` convention plugin in
a new included `build-logic` build, applied per-project. build-logic
carries ktfmt and the Kotlin Gradle plugin on its classpath so ktfmt can
decorate against KotlinSourceSet; because it's an included build (not
buildSrc) with versions from the shared catalog, modules keep their
existing `alias(...)` plugins without classpath clashes. The AGP-9 ktfmt
source-set workaround for Android modules now lives once in that plugin
instead of being repeated per module. This also leaves the build ready for
Isolated Projects, which disallows the old cross-project `allprojects {}`
form.

Deprecations:
- app: replace the removed Kotlin `srcDir(Any)` with `directories.add(...)`.
- CardHistoryScreen: use the AutoMirrored AddToHomeScreen icon.

Isolated Projects itself is deferred: enabling it in gradle.properties
breaks the compose-preview tooling (CLI/MCP/CI render), whose auto-inject
init script configures projects via `allprojects {}`. Revisit once that
init script becomes IP-compatible upstream.
