+++ rc-converter/build.gradle.kts
@@ -1,4 +1,5 @@
 plugins {
+  id("harc.base-conventions")
   alias(libs.plugins.android.library)
   alias(libs.plugins.compose.compiler)
 }
